+++ src/GUI/UI/Houses/Bank.cpp
@@ -44,7 +44,7 @@ void GUIWindow_Bank::putGoldDialogue() {
         return;
     }
     if (keyboard_input_status == WINDOW_INPUT_CONFIRMED) {
-        int sum = static_cast<int>(std::strtol(keyboardInputHandler->GetTextInput().c_str(), nullptr, 10));
+        int sum = std::strtol(keyboardInputHandler->GetTextInput().c_str(), nullptr, 10);
         if (sum <= 0) {
             engine->_messageQueue->addMessageCurrentFrame(UIMSG_Escape, 1, 0);
             return;
@@ -84,7 +84,7 @@ void GUIWindow_Bank::getGoldDialogue() {
         return;
     } else if (keyboard_input_status == WINDOW_INPUT_CONFIRMED) {
         keyboard_input_status = WINDOW_INPUT_NONE;
-        int sum = static_cast<int>(std::strtol(keyboardInputHandler->GetTextInput().c_str(), nullptr, 10));
+        int sum = std::strtol(keyboardInputHandler->GetTextInput().c_str(), nullptr, 10);
         if (sum <= 0) {
             engine->_messageQueue->addMessageCurrentFrame(UIMSG_Escape, 1, 0);
             return;
